--- src/components/About/Awards.module.css
@@ -1,15 +1,0 @@
-.cv-box {
-  padding: var(--padding-even-md);
-  display: flex;
-  flex-flow: row wrap;
-}
-
-.award-box {
-  flex: 1 1 calc(50% - 0.5rem);
-}
-
-@media all and (max-width: 768px) {
-  .award-box {
-    flex: 1 1 100%;
-  }
-}
