--- src/components/About/Involvement.module.css
@@ -1,4 +1,0 @@
-.text {
-  padding: var(--padding-even-md);
-  margin: 0;
-}
