--- src/components/About/Kompetenzen.module.css
@@ -1,92 +1,0 @@
-.tab-box {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 2rem;
-  padding: 1rem;
-  width: 100%;
-}
-
-.elements {
-  flex: 1 0 calc(50% - 2rem);
-}
-
-@media all and (max-width: 768px) {
-  .elements {
-    flex: 1 0 100%;
-  }
-}
-
-.title {
-  font-size: 1rem;
-  font-weight: 500;
-  letter-spacing: 0.5px;
-  position: relative;
-  overflow: hidden;
-}
-
-.title::first-letter {
-  font-size: 2.5rem;
-  font-weight: 500;
-  margin-right: -0.2rem;
-  padding: 0.3rem;
-  color: var(--color-primary);
-}
-
-.title:after {
-  position: absolute;
-  background: var(--color-primary);
-  content: '';
-  height: 0.25rem;
-  top: 0;
-  width: 1260px;
-}
-
-.title:after {
-  margin-left: 1rem;
-}
-
-.skill-box {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.skill--logo,
-.skill--text {
-  flex: 1 1 auto;
-  display: flex;
-  align-items: center;
-  padding: var(--padding-even-sm);
-  border: 1px solid var(--color-bg-transparent60);
-  border-radius: var(--border-radius);
-}
-
-.skill--logo {
-  background-color: var(--color-bg-transparent20);
-}
-
-.skill--text {
-  font-size: 0.9rem;
-  background-color: var(--color-bg-transparent20);
-}
-
-.skill--logo:last-child,
-.skill--text:last-child {
-  flex: 0 0 auto;
-}
-
-.logo {
-  height: 1.8rem;
-  width: auto;
-}
-
-.name {
-  font-size: 0.9rem;
-  padding: var(--padding-even-sm);
-  font-weight: 400;
-}
-
-.customize-tags {
-  justify-content: center;
-}
