--- src/components/About/Work.module.css
@@ -1,48 +1,0 @@
-.cv-box {
-  padding: var(--padding-even-md);
-  display: flex;
-  flex-flow: row wrap;
-}
-
-.work-box {
-  flex: 1 1 calc(50% - 0.5rem);
-}
-
-.work-box:nth-child(1) {
-  order: 1;
-}
-
-.work-box:nth-child(2) {
-  order: 3;
-}
-
-.work-box:nth-child(3) {
-  order: 5;
-}
-
-.work-box:nth-child(4) {
-  order: 2;
-}
-
-.work-box:nth-child(5) {
-  order: 4;
-}
-
-.work-box:nth-child(6) {
-  order: 6;
-}
-
-@media all and (max-width: 768px) {
-  .work-box {
-    flex: 1 1 100%;
-  }
-
-  .work-box:nth-child(1),
-  .work-box:nth-child(2),
-  .work-box:nth-child(3),
-  .work-box:nth-child(4),
-  .work-box:nth-child(5),
-  .work-box:nth-child(6) {
-    order: 1;
-  }
-}
